add route and integration tests
