+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/integration_tests/tests/web/UsersTests.kt
@@ -4,6 +4,8 @@ import org.junit.Test
 import org.vaccineimpact.orderlyweb.ContentTypes
 import org.vaccineimpact.orderlyweb.models.Scope
 import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
+import org.vaccineimpact.orderlyweb.tests.addMembers
+import org.vaccineimpact.orderlyweb.tests.createGroup
 import org.vaccineimpact.orderlyweb.tests.integration_tests.tests.IntegrationTest
 import spark.route.HttpMethod
 
@@ -26,6 +28,29 @@ class UsersTests : IntegrationTest()
         assertWebUrlSecured(url, setOf(ReifiedPermission("users.manage", Scope.Global())),
                 contentType = ContentTypes.json, method = HttpMethod.post, postData = mapOf("action" to "add",
                 "name" to "users.manage"))
+    }
+
+    @Test
+    fun `only user managers can get global report reading groups`()
+    {
+        val url = "/user-groups/report-readers/"
+
+        assertWebUrlSecured(url, setOf(ReifiedPermission("users.manage", Scope.Global())),
+                contentType = ContentTypes.json)
+    }
+
+    @Test
+    fun `can get global report reading groups`()
+    {
+        createGroup("Funder", ReifiedPermission("reports.read", Scope.Global()))
+        addMembers("Funder", "funder.a@example.com", "funder.b@example.com")
+
+        val url = "/user-groups/report-readers/"
+        val response = webRequestHelper.loginWithMontaguAndMakeRequest(url,
+                setOf(ReifiedPermission("users.manage", Scope.Global())),
+                ContentTypes.json)
+
+        JSONValidator.validateAgainstSchema(response.text, "UserGroups")
     }
 
 }
