+++ src/testHelpers/src/main/kotlin/org/vaccineimpact/orderlyweb/test_helpers/insertHelpers.kt
@@ -147,4 +147,3 @@ fun giveUserGlobalPermission(db: JooqContext, userGroup: String, permissionName:
         this.id = id
     }.insert()
 }
-
